feat(selected): 주문 시 선택 데이터 삭제
